pyproject.toml
[project]
name = "fastapi-owasp-api-security-top10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.135.1",
    "loguru>=0.7.3",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "ruff>=0.15.5",
    "sqlmodel>=0.0.37",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
  "F","E","W","I","UP","B","C4","SIM","RET","ARG","N","D","ANN","ASYNC"
]
ignore = [
  "ANN001","ANN002","ANN101","ANN102",
  "E501" # formatter handles line length
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104","F401"]


[tool.mypy]
python_version = "3.13"
files = ["app/**/*.py"]
strict = true
disallow_any_unimported = true
warn_unreachable = true
no_implicit_optional = true
show_error_codes = true
ignore_missing_imports = true
